+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## [1.1.0-beta.0](https://github.com/Seungwoo321/genai-commit/compare/v1.0.2...v1.1.0-beta.0) (2026-01-17)
+
+### Features
+
+* add error reporting link on commit generation failure ([aaeab3f](https://github.com/Seungwoo321/genai-commit/commit/aaeab3f6312986fdee446d66e1b811f5329d5b28))
+* add interactive mode support for CLI commands ([e081ae7](https://github.com/Seungwoo321/genai-commit/commit/e081ae74733f5ca74766801dcaa19f7fdc139554))
+* add models command to list supported models ([ab05bbb](https://github.com/Seungwoo321/genai-commit/commit/ab05bbbb5e9e6654f09ff46a212ce789c07d32ab))
+* update default Cursor model and add model lists ([5ca52bc](https://github.com/Seungwoo321/genai-commit/commit/5ca52bc8f9fcabb537be0bb13757defb10fb8636))
+
+### Bug Fixes
+
+* correct Cursor CLI command invocation ([9f92b44](https://github.com/Seungwoo321/genai-commit/commit/9f92b44f510d0b6d808a87aef02c92bcbaba6fa9))
+
+### Documentation
+
+* update CLI examples and add models command ([7e27e6b](https://github.com/Seungwoo321/genai-commit/commit/7e27e6b5bd09c45c161e8f9443a1d60516189cd9))
+
 ## [1.0.2](https://github.com/Seungwoo321/genai-commit/compare/v1.0.1...v1.0.2) (2026-01-17)
 
 ### Bug Fixes